move files to definition files where possible
