+++ packages/back-end/src/events/base-types.ts
@@ -1,3 +1,5 @@
+import { EventAuditUser } from "./event-types";
+
 /**
  * Supported events for event notifications
  */
@@ -39,36 +41,3 @@ export type NotificationEventPayload<
   data: DataType;
   user: EventAuditUser;
 };
-
-// region Audit
-
-/**
- * You can get this property on the response.locals.eventAudit property
- */
-export type EventAuditUser =
-  | EventAuditUserLoggedIn
-  | EventAuditUserApiKey
-  | null;
-
-/**
- * You can get this property on the response.locals.eventAudit property.
- * Example usage:
- *    (req, res: Response<MyResponseData, EventAuditUserForResponseLocals>) => {}
- */
-export type EventAuditUserForResponseLocals = {
-  eventAudit: EventAuditUser;
-};
-
-export type EventAuditUserLoggedIn = {
-  type: "dashboard";
-  id: string;
-  email: string;
-  name: string;
-};
-
-export type EventAuditUserApiKey = {
-  type: "api_key";
-  apiKey: string;
-};
-
-// endregion Audit
