+++ packages/back-end/src/events/event-types.d.ts
@@ -1,0 +1,46 @@
+// region user.login
+
+export type UserLoginAuditableProperties = {
+  id: string;
+  email: string;
+  name: string;
+  device: string;
+  userAgent: string;
+  ip: string;
+  os: string;
+};
+
+// endregion user.login
+
+// region Audit
+
+/**
+ * You can get this property on the response.locals.eventAudit property
+ */
+export type EventAuditUser =
+  | EventAuditUserLoggedIn
+  | EventAuditUserApiKey
+  | null;
+
+/**
+ * You can get this property on the response.locals.eventAudit property.
+ * Example usage:
+ *    (req, res: Response<MyResponseData, EventAuditUserForResponseLocals>) => {}
+ */
+export type EventAuditUserForResponseLocals = {
+  eventAudit: EventAuditUser;
+};
+
+export type EventAuditUserLoggedIn = {
+  type: "dashboard";
+  id: string;
+  email: string;
+  name: string;
+};
+
+export type EventAuditUserApiKey = {
+  type: "api_key";
+  apiKey: string;
+};
+
+// endregion Audit
