+++ packages/back-end/src/events/notification-events.d.ts
@@ -1,6 +1,6 @@
 import { ApiExperiment, ApiFeature } from "../../types/openapi";
 import { NotificationEventPayload } from "./base-types";
-import { AuditableUserProperties } from "./event-types";
+import { UserLoginAuditableProperties } from "./event-types";
 
 // region User
 
@@ -8,7 +8,7 @@ export type UserLoginNotificationEvent = NotificationEventPayload<
   "user.login",
   "user",
   {
-    current: AuditableUserProperties;
+    current: UserLoginAuditableProperties;
   }
 >;
