+++ packages/back-end/src/services/auth/index.ts
@@ -23,7 +23,7 @@ import { getPermissionsByRole } from "../../util/organization.util";
 import {
   EventAuditUserForResponseLocals,
   EventAuditUserLoggedIn,
-} from "../../events/base-types";
+} from "../../events/event-types";
 import { AuthConnection } from "./AuthConnection";
 import { OpenIdAuthConnection } from "./OpenIdAuthConnection";
 import { LocalAuthConnection } from "./LocalAuthConnection";
